page: Set rev id in OutputPage from dirty ParserOutput on action=view
In case a dirty ParserOutput must be used for the page view,
also copy over the revision id to the OutputPage to produce consistent
html.
The already set revision id is from fresh replica (or primary),
but the html is not ready to render yet and old html is used.

Bug: T341013
Change-Id: Iaa558da8676c66296fe0f092857db82ed2389c1c