ResourceLoader: Simplify clientpref storage
After this change, for the 99% case, all we do is:

* read local string variable (doc.cookie),
* match a regex that has no matches,
* 1 assignment to docEl.className.

This are the same three operations as before change I1e635f843ac.
Specifically, this avoids:
* build a data structure from the cookie.
* match and extract every feature class (even when unchanged),
* perform a no-op replacement for each one,

For the case where client prefs do exist, this reduces the logic to
a string replacement (not a DOM read-write), and doing so only N times
(for each customisation), instead of the maximum numer of times for
every feature that could be customised.

Bug: T339268
Change-Id: Ia85ef5c2be7bf759bc6041f8a999eeb584e554e8