ooui: SearchInputWidget must send request start events
The autocomplete metrics collection, which collects from three
separate javascript implementations, needs to receive the same
events from each of the implementations. On review it was found
that the other implementations send this session-start event prior
to each backend request, and the metrics collection code uses that
to record user perceived autocomplete latency.

Update ooui SearchInputWidget to send the same session-start event
that the other implementations do.

Bug: T335499
Change-Id: Ibfb270337b1cec02c039dcb7238bc3a559bd1fff