SpecialPage::getDescription() should return a Message
Deprecate returning a string from SpecialPage::getDescription(), which
encourages good localization as well as helping us shift from
OutputPage::setPageTitle() to ::setPageTitleMsg(), which in turn helps
improve the accuracy of phan-taint-check.

In the process, special page description messages are formatted using
FORMAT_ESCAPED by OutputPage::setPageTitleMsg(), which is a change from
FORMAT_TEXT used prior to this patch.  Descriptions shouldn't contain
raw HTML, but if a description message previously contained an HTML
metacharacter like <, >, or &, it will no longer need to be escaped
in the message localization.

Bug: T343994
Bug: T343849
Depends-On: I1a27f7697dffd3142cdbab0c8cd9854b2b94dd3a
Change-Id: Ifc3452548c44a021414cf8433c1ffbb683a9d55b