rdbms: Avoid making a db query when getting MySQL/MariaDB version
Currently, DatabaseMySQLBase::selectSQLText() needs to modify the SQL
sometimes and in those cases, it needs to get the MySQL/MariaDB version.

To get the version, you don't actually need to call "SELECT VERSION()"

The PDO provides that value for free and from reading the source code,
It's provided during the handshake:
https://github.com/php/php-src/blob/cb927e0fc043cb454ee58b7485cbce191df5d512/ext/mysqlnd/mysqlnd.h#L136

The only problem is that some systems, prefix it with "5.5.5-"
(including my localhost) so let's just strip that instead of making
extra DB query.

Change-Id: I665d6b8583632a09e25efea976875acc7e484cee