http: fix JobRunner overriding requestId with null
After deploying T320559 we found out that JobRunner sometimes calls
the WebRequest::overrideRequestId() with a null. This caused the
TelemetryService to fail due to $newId not being a string.

As a solution there is a new method called regenerateRequestId()
that will handle the regeneration as it's recommended to keep the
overrideRequestId() strict.

Bug: T343893
Change-Id: I3e6daa238fd8de0d9ea88660b8ba32b922b1323d