Deprecate returning a string from Action::getPageTitle()
In almost all cases we're already returning a Message object from
Action::getPageTitle().  Tighten this up by deprecating the string
return value and requiring the returned value be a Message; this also
encourages proper localization.

We pass the returned message to OutputPage::setPageTitleMsg(), which
uses FORMAT_ESCAPED.  This is a change from before this patch, where
messages are formatted using FORMAT_TEXT.  If you need to include
HTML in the page title, use Message::rawParams().

Bug: T343994
Bug: T343849
Depends-On: I2fe88d95737e673a381d039909e854e4410c8ce3
Change-Id: Ia7da1b7063a77bf94f4c71e9d77c9408c3f4aa46