Make MediaWikiIntegrationTestCase::needsDB() final and static
Subclasses are not supposed to override this to change the behaviour.
Any test that needs the database, even if conditionally, should be in
the Database group to allow filtering.

Also remove the $tablesUsed check: we already throw an exception if the
test has $tablesUsed but is not in the Database group, so there's no
need to check that twice. This also allows us to make the method static.
Move said check to the setup method, as that's where all the setup
should be eventually.

Update the documentation on various methods to reflect the fact that the
only way to mark a test as using the database is via the Database group.

Note that making a method static is a backwards-compatible change, but
callers in MediaWikiIntegrationTestCase are updated anyway for
consistency.

Bug: T342259
Depends-On: Ied6a69a1e4beb7fbb4d5ce75b57f54382d929684
Change-Id: Ie72a0dff902feee92a183b2dc2926bd2ad1df071