Do not override run() in MediaWikiIntegrationTestCase
This has always been a hack. PHPUnit provides hook methods exactly for
this. It might have been a bit better 13 years ago [1], when using
PHPUnit 3 and this method only had 3 lines, but today it's become
bloated.

PHPUnit 10 makes run() final, so we'll have to move away sooner or
later.

Move code to the @before and @after methods, which are run relatively
close to run().

Make MediaWikiIntegrationTestCaseTest use @depends to test that the
tearDown logic works correctly. The previous code was too tightly
coupled with the specific implementation of mediaWikiTearDown.

[1] - https://static-codereview.wikimedia.org/MediaWiki/79173.html

Bug: T342259
Change-Id: I4dee53176fc90a7884600422c30dc904cbd1a7d6