Make MediaWikiIntegrationTestCase::resetDB static
The method no longer has non-static dependencies. By making it static,
it will be possible to call it in the tearDownAfterClass hook method to
clear tables filled in addDBDataOnce.

Also make the IDatabase parameter not nullable, none of the two callers
is ever passing null. And also make the method return early if
$tablesUsed is empty.

Change-Id: I4bac154cf6bb4007abd4be7ca31090de5d33e7d5