watchlist: Prepare for IP Masking
Instead of replacing isRegistered with isNamed
in the watchlist component, (re)assign the watchlist
rights (`editmywatchlist` and `viewmywatchlist`) to
the `user` group, instead of `*`. This should hide
watchlist-related interface from temp accounts, as
those are not a member of the `user` group.

isRegistered() is kept in watchlist, as technically,
any user with an ID could have a watchlist (it is now
a matter of rights / product decision, rather than
technological reasons).

This patch partially reverts
Ib935fc04387100f25cafd75c632ac5ae7114d5d5.

This is an alternative to Ifeedbbfb520ec26680134279bd44fec98754d897.

What is left is to change the "User registration"
filters to match expectations for temp. accounts
(to be done in a follow-up patch; see T343322).

Bug: T341976
Change-Id: I1ccda3878a658d4251429ae65ebee04c09d81243