ContribsPager: Put tag parsing behind an in-process cache
Apply the idea promoted in a468534 to Special:Contributions.
If a user uses an external tool (or just the visual editor),
they are likely to do it repeatedly during a period of time.
In turn, their edits will be tagged with the same set of tags.
Instead of parsing the tags from scratch again and again,
memoize the result for use by subsequent calls.

Thanks to WikimediaDebug, I could reproduce a request to
Special:Contributions?limit=500 where 28% (1.6s) of request
time was spent in ChangeTags::formatSummaryRow. This change
should reduce it to less than 0.1s.

Change-Id: If81d7b5359f798f0b630ab1ad5b6282d0d9d9a55