WANObjectCache: don't set a hold-off when the cache is empty
When getWithSetCallback() is called with check keys, if the keys are
missing, a check key is inserted with the current time, as if
touchCheckKey() were called. This causes cache misses for
HOLDOFF_TTL = 11 seconds. This seems unnecessary since in the case of
an empty cache, there is no expectation of replication delay.

However, it's reasonable for it to be a cache miss when the check key is
missing, and a cache hit subsequently, so we do need to add a purge
value.

So, in getWithSetCallback(), set the holdoff to zero when inserting a
purge value.

Also, use a holdoff of zero when initialising a missing touch key in
getCheckKeyTime().

Bug: T344191
Change-Id: Ib3ae4b963816e5b090e87e4cb93624afefbf8058