deferred: Move DeferredUpdates away from DUManager service approach
DU is a PHP process level mechanism to finish things related to a MW
request (which is possible to context changes). By request, meaning it's
stateful and a lot of things can be altered before the response gets to
the client or completes.

Services on the other hand are by design stateless meaning that making
DU stateless is approaching things from a contrary perspective. This means
when an update is deferred, nothing can change anymore as it's now done
so via a DUManager but that's the opposite of what we want becuase an
update (whether presend or postsend) should be able to do other things
in the background (perform changes to the DB, change an article's rendering
etc).

This patch is an attempt to move DU away from being a service class.

Bug: T265749
Depends-On: I2b0e80bdd92a5fd4d7c325fc8792df45e0776c30
Depends-On: I8126cc76440724753c356c48ba4e0fcc9be5b41a
Depends-On: Ia2fa060237c4e9d328d1b2b9622a65fcca97b854
Change-Id: Ifbfb0442bf6264d0bf786334416c5d99c0dd4699