cross-wiki userrights: Restore hook compatibility
In c87c8f64 (I2827e99d09) the user argument of the hooks
"UserGroupsChanged" and "ChangeUserGroups" were changed to always be a
User object.

Previously it was either a User object (for local user group
changes) or a UserRightsProxy object (for interwiki group changes), but
documented it as User object by 493e2de5 (I0c56b6f2b7), which makes the
wrong assumption to change it to always a User object without
compatibility issues.

But with User objects it is not possible for the hook handler to detect
the interwiki group change from local one, Echo want to skip interwiki
group changes as it cannot insert the notification on the target wiki.

Document the different behaviour in the hooks.
For full compatibility this must be UserRightsProxy, but that is
deprecated and was replaced, so not an option.
Changing the hook to UserIdentity does not work as some extensions are
type hinted the hook handler for User (which would fail for interwiki
changes also before c87c8f64, as UserRightsProxy class is not compatible
with User class)

Bug: T344446
Bug: T255309
Follow-Up: I2827e99d0990fefee260c7e8dc272312a3aea7cb
Change-Id: Ia1e3cdc19f1ef9f6edeaa7bfe5a12669dd0b0ba2