Make OutputPageTest explicitly use English as content language
Some tests in this class expect the language to be English when checking
some generated HTML snippets.

Don't let the test be affected by LocalSettings, and set the content
language to English explicitly.

Switching to qqx is not immediately possible due to other assumptions
(e.g., that the language used has variants), but it would be nice for a
future change.

Bug: T234141
Change-Id: I91e13388bbfbcdae1e87e08a8676b56711bac855