Allow watchlist UI for temp users with sufficient rights
Following on from I1ccda3878a, if temp users are explicitly granted
viewmywatchlist and editmywatchlist rights in LocalSettings.php, make
the watchlist API and UI be functional for them.

The reasons for denying temp users access to the watchlist are mostly
Wikimedia-specific, and may change in future, so this provides a bit
more flexibility for third parties and for future use.

T330518 (Minerva watch star suppression) does not appear to regress.
This change does not cause a watch star to appear for Minerva, since it
is doing its own thing. A watch star does appear for Vector, conditional
on user rights.

Bug: T341976
Change-Id: Idfebd0f9ab939faa962dfad6d6f502c312c819fa