wdio-mediawiki: await saveScreenshot
On a test failure, I noticed it lacked the final screenshot as if the
content of our `afterTest` hook did not run. `browser.saveScreenshot` is
asynchronous and I guess it does not have an opportunity to
run/complete.

From https://github.com/webdriverio/webdriverio/issues/5545 and
specially the following comment:

> @christian-bromann wrote:
> > Aren't the hooks async by default?
>
> Sure, but if you do an async operation in there like saveScreenshot
> you have to define it as async function.

We thus have to await `browser.saveScreenshot` and explicitly define the
hook as being asynchronous so it can be waited for.

Bug: T337463
Change-Id: Icd38fdc48e3d280a901b341e99e394ba52678c84