Remove CHANGEDORCREATED from email templates
It was deprecated in 2013, and
"scheduled for removal in MediaWiki 1.23",
but never actually removed.

Now this parameter doesn't seem to be used anywhere
in the core code or in translations, so it, as well
as the associated messages, can be removed.

Follow-up to
Ibb7953741e4189127d1cd6718ac58492f254671e
and
I7d9fd90519a68dd66f99479a7c92f6c664485a9c

Change-Id: I1e8eed2f5b4421b0d387c27c9714b9ed7272a481