page: Run PageUndeleteComplete hook only on undeletion of text
It is possible to just undelete a file version,
in that case the restored revision is unknown/null and
running the hook fails with a TypeError.
As no new page gets created, the hook should not run.

Bug: T345018
Follow-Up: Id77d945dfe14f196c34652b4f3d2f3ec843cbe7c
Change-Id: I585ae4ca6eadd69c74be487a0e5e5147bef1959c