CalendarWidget: Make clippable, scrollable
This change adds the ClippableElement mixin to CalendarWidget,
making the element clippable and scrollable. It also makes
the calendar popup automatically flip (from opening downwards
to upwards, and vice versa) whenever there is not enough
space in the direction that the calendar opens.

This fixes an issue where the calendar is clipped by the
viewport when opening in a direction close to the edge of
the screen (e.g. when the DateInputWidget spawning the
CalendarWidget is close to the bottom of the screen and the
widget opens downwards).

Bug: T115939
Change-Id: I233de898618391e20a21bdf9c2e9639e88f71fb4