http: Allow MultiHttpClient handle headers on construction phase
HTTPFactory or any other class that initialises the MultiHttpClient
class should be able to pass a set of default/standard headers
that should be attached with every request.

Additionally, I noticed that passing headers do not work at
all when `curl` is not available. To pass headers to MwHttpClient
we need to call `$client->setHeader()`. Previously MultiHttpClient
passed them as $reqOptions['headers'], but this index wasn't checked
evaluated when performing the request.

Bug: T345295
Change-Id: I438cc2deab19a46668f8dcff5451d6607b1476a7