Don't set AuthenticationRequest::$username on login
When the user is already logged in, and attempts but fails to log in
again, the username in the PasswordAuthenticationRequest is the login
target, but the username in the RememberMeAuthenticationRequest is the
currently logged in user. The conflict causes $guessUserName to be set
to null. When the AuthManagerLoginAuthenticateAudit hook is called in
the fail case, the User is not sent, only $guessUserName, so LoginNotify
has no way to notify the target user of the failed attempt.

So, don't populate AuthenticationRequest::$username with the currently
logged-in username on login, by analogy with create account. Then
LoginNotify works as expected.

Are there unanticipated consequences? I think it should be generally
safe to treat the user as not logged in when they attempt to log in as
another user. ThrottlePreAuthenticationProvider is apparently trying to
use the target username for throttling, not the currently logged-in
user, which seems like a good policy.

I couldn't find a rationale in Gerrit for populating the username on
login. In PS72 of If89d24838e326fe25fe867d02, Gergő asked for
documentation, but not a rationale. It was convenient to treat login the
same as link/unlink/change. The lack of a test case suggests it was not
intentional.

Bug: T329774
Change-Id: Id11f3da62a7f4fb3571a605c62a3924d6a8ee50c