objectcache: Use DBO_DEBUG when DebugDumpSql setting is set to `true`
This is to be consistent with the DBO_* constants in IDatabaseFlags.php
not use the setting value directly.

This worked in the past because it evaluates to true or false based on
$sqlDump, which means instead of setting flags to DBO_DEBUG (int), it was
setting it to bool(true).

This worked by accident because:
* when we perform -, +, | or & math on a value, it is casted to an integer,
* and intval(true) === 1,
* and DBO_DEBUG is the first constant in DatabaseFlagsHolder, which we gave
  the number 1.

It is only the combination of all three of these facts that made it work by
accident.

Follow-up: I4ce691b77f775c20bcf4e8deb2ec1aae1b4674d8
Bug: T318272
Change-Id: I8ba1d35f0e36ae259830fc1f65fcb4e4dc92a8ec