maintenance: Move `shutdown()` logic to MaintenanceRunner
The shutdown logic intuitively feels like the job of the runner and
the maintenance script themselves don't need to have knowledge about
this routine.

Maintenance script should really only execute their tasks after being
invoked by the runner and leave the runner to clean the house at the
end of their execution. Hence shifting the logic to the runner class.

Change-Id: I3d6fbf02b5fcd1414fde3a84e98a08e58456d668