rdbms: Use `debugSql` instead of `debugDumpSql` which is unuset
We introduced new central logic to log SQL queries when the debug
flag for SQL queries is set to true.

But in Ia89de3b32b4db1b7abc9ed3ac676af4e7ac18ffa when we introduced
the new option, we referenced it with the wrong name. This patch
fixes the index reference.

Bug: T318272
Change-Id: I112f25c53e7b257230d9d6c4cc8d23d513f0f55b