skins: Add a data-mw-expiry attribute to watch link
When a page is only watched temporary, set the iso timestamp on the
watch link to allow user scripts using the expiry without asking via api
query
This also updates the timestamp when the watch status is changed on the
page, this timestamp is the iso timestamp with microseconds, while php
timestamp is without microseconds. JavaScript's Date can handle both.

Bug: T270057
Change-Id: I4abf2daae024e051fbaa3db9a9ecfa92d7bf265f