Remember selected watchlist expiry time during edit conflict
Steps to reproduce (details in T274658):
* Edit a page in classic wikitext mode.
* Choose a new expiry time at the bottom of the form.
* Submit.
* When you run into a conflict now, the form is shown again, but the
  expiry time you just picked is gone.

Note I'm not 100% sure if this is the best possibly fix. However, it
appears to do exactly what it should do in my local experiments.

By the way, can we please stop piling up code in this horrendous
class? It's 4600 lines long by now, with up to 500 lines in a single
method.

Bug: T274658
Change-Id: I4ef7c23ddf96e963219714944467bd28fe47acba